# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Data Bundle Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bundles:
    get:
      operationId: listBundles
      tags: [bundles]
      summary: List all active data bundles
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bundle"

    post:
      operationId: createBundle
      tags: [bundles]
      summary: Create a new bundle (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBundleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bundles/{id}:
    get:
      operationId: getBundle
      tags: [bundles]
      summary: Get a bundle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateBundle
      tags: [bundles]
      summary: Update a bundle (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBundleBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteBundle
      tags: [bundles]
      summary: Delete a bundle (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: List orders for the current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Purchase a bundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Bundle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        dataAmount:
          type: string
        validityDays:
          type: integer
        price:
          type: number
        category:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - dataAmount
        - validityDays
        - price
        - category
        - isActive
        - createdAt
    CreateBundleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        dataAmount:
          type: string
        validityDays:
          type: integer
        price:
          type: number
        category:
          type: string
      required:
        - name
        - description
        - dataAmount
        - validityDays
        - price
        - category
    UpdateBundleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        dataAmount:
          type: string
        validityDays:
          type: integer
        price:
          type: number
        category:
          type: string
        isActive:
          type: boolean
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bundleId:
          type: integer
        bundleName:
          type: string
        bundleData:
          type: string
        price:
          type: number
        status:
          type: string
        phoneNumber:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - bundleId
        - bundleName
        - bundleData
        - price
        - status
        - phoneNumber
        - createdAt
    CreateOrderBody:
      type: object
      properties:
        bundleId:
          type: integer
        phoneNumber:
          type: string
      required:
        - bundleId
        - phoneNumber
...